requirements/requirements-base.txt
python-dotenv
jupyterlab-vim
nltk
scikit-learn
pandas
seaborn
beakerx
gensim
xlrd
ipdb
datasets
tqdm
